Update [profile-plugin]: finish renaming
